+++ frontend/src/styles/style.scss
@@ -1,3 +1,7 @@
+:root {
+  --header-height: 6vh;
+}
+
 html {
   background: #e5e5e5;
 }
@@ -10,4 +14,11 @@ body {
     "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
     "Segoe UI Symbol", "Noto Color Emoji";
   background: white;
+}
+
+main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: calc(100vh - var(--header-height));
 }
